requirements.txt
openai-whisper
gtts
moviepy
torch 
python-dotenv
requests 